pyproject.toml
[project]
name = "docs-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-contentunderstanding>=1.0.1",
    "azure-identity>=1.25.3",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mistralai>=2.1.3",
    "openai>=2.30.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
